src/edge/simulation/pyproject.toml
[project]
name = "simulation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "geojson>=3.2.0",
    "matplotlib>=3.10.8",
    "minio>=7.2.20",
    "numpy>=2.4.0",
    "paho-mqtt>=2.1.0",
    "scipy>=1.16.3",
]
